Modified methods to be user specific

+++ Tweetbook/Controllers/V1/PostsController.cs
@@ -8,6 +8,7 @@ using Tweetbook.Contracts.V1.Requests;
 using Tweetbook.Contracts.V1.Responses;
 using Tweetbook.Domain;
 using Tweetbook.Services;
+using Tweetbook.Extensions;
 
 namespace Tweetbook.Controllers.V1
 {
@@ -41,6 +42,11 @@ namespace Tweetbook.Controllers.V1
         [HttpPut(ApiRoutes.Posts.Update)]
         public async Task<IActionResult> Update([FromRoute] Guid postId, [FromBody] UpdatePostRequest request)
         {
+            var userOwnsPost = await _postService.UserOwnsPost(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+                return BadRequest(new { error = "You don't own this post" });
+
             var post = new Post
             {
                 Id = postId,
@@ -54,6 +60,11 @@ namespace Tweetbook.Controllers.V1
         [HttpDelete(ApiRoutes.Posts.Delete)]
         public async Task<IActionResult> Delete([FromRoute] Guid postId)
         {
+            var userOwnsPost = await _postService.UserOwnsPost(postId, HttpContext.GetUserId());
+
+            if (!userOwnsPost)
+                return BadRequest(new { error = "You don't own this post" });
+
             var deleted = await _postService.DeletePostAsync(postId);
 
             if (deleted)
@@ -65,15 +76,22 @@ namespace Tweetbook.Controllers.V1
         [HttpPost(ApiRoutes.Posts.Create)]
         public async Task<IActionResult> Create([FromBody] CreatePostRequest postRequest)
         {
-            var post = new Post { Name = postRequest.Name };
+            var post = new Post
+            {
+                UserId = HttpContext.GetUserId(),
+                Name = postRequest.Name
+            };
 
-            await _postService.CreatePostAsync(post);
+            await _postService.CreatePostAsync(post); 
+            var response = new PostResponse { Id = post.Id };
+            return Created(UriLocation(post), response);
+        }
 
+        private string UriLocation(Post post)
+        {
             var baseUrl = HttpContext.Request.Scheme + "://" + HttpContext.Request.Host.ToUriComponent();
             var uriLocation = baseUrl + "/" + ApiRoutes.Posts.Get.Replace("{postId}", post.Id.ToString());
-
-            var response = new PostResponse { Id = post.Id };
-            return Created(uriLocation, response);
+            return uriLocation;
         }
     }
 }
